backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.6.0
pytz>=2024.1

asyncpg>=0.29.0

# --- pinned for Render stability ---
passlib==1.7.4
bcrypt==4.0.1

# NOTE: Do not add emergentintegrations - we use local stripe_checkout and no Emergent LLM/Auth.

# --- Added by Render Sledgehammer ---
twilio>=9.0.0
apscheduler>=3.10.0
python-jose[cryptography]>=3.3.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
vobject>=0.9.6
python-multipart>=0.0.9
stripe>=8.0.0
